+++ testcontainers/src/core/image.rs
@@ -27,13 +27,25 @@ where
     /// testing. Only expose those that actually make sense for this case.
     type Args;
 
-    /// The name of the docker image to pull from the Docker Hub registry.
+    /// The name of the image to pull from a registry.
     fn name(&self) -> String;
 
     /// Implementations are encouraged to include a tag that will not change (i.e. NOT latest)
     /// in order to prevent test code from randomly breaking because the underlying docker
     /// suddenly changed.
     fn tag(&self) -> String;
+
+    /// The registry to use for the image descriptor. Optional, by default it uses DockerHub.
+    /// Will be prepended to image descriptor if present. E.g: `{registry}/{owner}/{name}`
+    fn registry(&self) -> Option<String> {
+        None
+    }
+
+    /// The owner of the image in a registry to use for the image descriptor.
+    /// Optional, will be prepended to image descriptor if present (before `registry`). E.g: `{registry}/{owner}/{name}`
+    fn owner(&self) -> Option<String> {
+        None
+    }
 
     /// Returns a list of conditions that need to be met before a started container is considered ready.
     ///
@@ -151,6 +163,8 @@ pub struct RunnableImage<I: Image> {
     image_args: I::Args,
     image_name: Option<String>,
     image_tag: Option<String>,
+    image_registry: Option<String>,
+    image_owner: Option<String>,
     container_name: Option<String>,
     network: Option<String>,
     env_vars: BTreeMap<String, String>,
@@ -203,13 +217,28 @@ impl<I: Image> RunnableImage<I> {
     }
 
     pub fn descriptor(&self) -> String {
+        let original_registry = self.image.registry();
+        let original_owner = self.image.owner();
         let original_name = self.image.name();
         let original_tag = self.image.tag();
+
+        let registry = self
+            .image_registry
+            .as_ref()
+            .or(original_registry.as_ref())
+            .map(|r| format!("{}/", r))
+            .unwrap_or_default();
+        let owner = self
+            .image_owner
+            .as_ref()
+            .or(original_owner.as_ref())
+            .map(|o| format!("{}/", o))
+            .unwrap_or_default();
 
         let name = self.image_name.as_ref().unwrap_or(&original_name);
         let tag = self.image_tag.as_ref().unwrap_or(&original_tag);
 
-        format!("{name}:{tag}")
+        format!("{registry}{owner}{name}:{tag}")
     }
 
     pub fn ready_conditions(&self) -> Vec<WaitFor> {
@@ -226,7 +255,6 @@ impl<I: Image> RunnableImage<I> {
 }
 
 impl<I: Image> RunnableImage<I> {
-    /// Overrides the image name. Can be used to specify a custom registry or owner (e.g `{domain}/{owner}/{image}`).
     pub fn with_name(self, name: impl Into<String>) -> Self {
         Self {
             image_name: Some(name.into()),
@@ -239,6 +267,20 @@ impl<I: Image> RunnableImage<I> {
     pub fn with_tag(self, tag: impl Into<String>) -> Self {
         Self {
             image_tag: Some(tag.into()),
+            ..self
+        }
+    }
+
+    pub fn with_registry(self, registry: impl Into<String>) -> Self {
+        Self {
+            image_registry: Some(registry.into()),
+            ..self
+        }
+    }
+
+    pub fn with_owner(self, owner: impl Into<String>) -> Self {
+        Self {
+            image_owner: Some(owner.into()),
             ..self
         }
     }
@@ -308,6 +350,8 @@ impl<I: Image> From<(I, I::Args)> for RunnableImage<I> {
             image_args,
             image_name: None,
             image_tag: None,
+            image_registry: None,
+            image_owner: None,
             container_name: None,
             network: None,
             env_vars: BTreeMap::default(),